move Future template in its own header
