+++ cvmfs/signing_tool.h
@@ -10,7 +10,7 @@
 
 #include "crypto/hash.h"
 #include "server_tool.h"
-#include "util/concurrency.h"
+#include "util/future.h"
 
 class ServerTool;
